requirements.txt
# Sistema de Administración - Supermercado
# Dependencias del proyecto

# Framework Web
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Base de datos
psycopg2-binary>=2.9.9

# Visualización
plotly>=5.18.0
pandas>=2.0.0

# API y HTTP
requests>=2.31.0

# Variables de entorno
python-dotenv>=1.0.0

# Utilidades
numpy>=1.24.0
